fix(agent-core): define subset closure relative to declared properties
CORE-039. proposal-reviewer ruled (REVISE -> applied, 2026-08-16) after its own
blast-radius claim was falsified during implementation: its repository sweep had
excluded test files, and a shipped test declares a bare nested `{ type: 'object' }`
and asserts a populated payload is accepted.

`additionalProperties` declares closure RELATIVE TO a declared `properties` set.
A node that declares `properties` -- including an empty one -- is closed unless
`additionalProperties` says otherwise; a node that declares none permits any
properties, as JSON Schema says and as every provider reads the document we
forward them. This is not an exception to the convention: the convention was
authored for a tool's `parameters` root, where `properties` is structurally
always present, and carrying its phrasing to a position that can omit the member
would silently give omitted `additionalProperties` a second meaning ("the empty
object only") that no producer intends -- the same root-versus-nested divergence
this item removes. Keyed on member presence, so a no-argument tool root still
rejects every argument, and implemented inside `validateAgainstJsonSchema` so
the two walks cannot diverge on it.

`packages/agent-tools/src/__tests__/function-tool.test.ts` is left untouched: it
encodes JSON Schema semantics correctly, and editing the one in-repo detector to
preserve the claim that there was nothing to detect is the wrong direction.

Splits two files that grew past their frozen size baselines rather than
extending them: the universal subset types move to
`packages/agent-core/src/interfaces/tool-schema.ts` (re-exported, so every
existing import resolves) and Anthropic's structured-output seam to
`packages/agent-provider-anthropic/src/anthropic/output-schema.ts`. The baseline
is re-tightened in the same change.

Naming the subset's object root as an interface cost the openai-compatible seam
its implicit index signature -- TypeScript grants one to an anonymous object type
but not to an interface -- so that seam widens by copy rather than by cast, which
keeps the conversion type-checked.
