+++ scripts/harness/file-size-baseline.json
@@ -12,7 +12,7 @@
   "packages/agent-core/src/core/robota.ts": 438,
   "packages/agent-core/src/index.ts": 350,
   "packages/agent-core/src/interfaces/agent.ts": 310,
-  "packages/agent-core/src/interfaces/provider.ts": 366,
+  "packages/agent-core/src/interfaces/provider.ts": 314,
   "packages/agent-core/src/managers/agent-factory.ts": 301,
   "packages/agent-core/src/plugins/event-emitter-plugin.ts": 310,
   "packages/agent-core/src/services/execution-round.ts": 308,
@@ -43,7 +43,7 @@
   "packages/agent-plugin/src/execution-analytics/execution-analytics-plugin.ts": 329,
   "packages/agent-plugin/src/usage/usage-plugin.ts": 301,
   "packages/agent-plugin/src/webhook/webhook-plugin.ts": 302,
-  "packages/agent-provider-anthropic/src/anthropic/provider.ts": 398,
+  "packages/agent-provider-anthropic/src/anthropic/provider.ts": 341,
   "packages/agent-provider-openai-compatible/src/deepseek/provider.ts": 302,
   "packages/agent-provider-openai-compatible/src/gemma/provider.ts": 312,
   "packages/agent-provider-openai-compatible/src/qwen/provider.ts": 301,
